+++ js/load.js
@@ -1,5 +1,11 @@
 window.onload = function() {
     console.log("Loading...");
+
     loadGame();
+
+    setInterval(function() {
+        saveGame();
+    }, 60000);
+
     console.log("Ready!");
 }